Create separate file for board_simulation class
